data type issue fixed

+++ TainaTech.Application/Features/Persons/Commands/UpdatePerson/UpdatePersonCommand.cs
@@ -6,7 +6,7 @@ namespace TainaTech.Application.Features.Persons.Commands.UpdatePerson
 {
     public class UpdatePersonCommand : IRequest
     {
-        public int PersonId { get; set; }
+        public long PersonId { get; set; }
         public string Firstname { get; set; }
         public string Surname { get; set; }
         public Gender Gender { get; set; }
